Move api.py back into core
