+++ minegauler/__init__.py
@@ -1,1 +1,2 @@
 from . import paths
+from .core import api
