+++ tests/mut/frontend/conftest.py
@@ -9,8 +9,7 @@ from unittest import mock
 
 import pytest
 
-from minegauler import api
-from minegauler.core import board
+from minegauler.core import api, board
 from minegauler.shared import utils
